feat: #7889 merged data-row-edit in de2
