+++ packages/data-explorer/package.json
@@ -29,43 +29,51 @@
     "@molgenis/molgenis-i18n-js": "^0.2.1",
     "@molgenis/molgenis-ui-context": "^1.6.5",
     "@molgenis/molgenis-ui-filter": "^1.11.1",
+    "@molgenis/molgenis-ui-form": "^5.2.2",
     "@molgenis/rsql": "^0.2.5",
     "axios": "^0.19.2",
     "bootstrap": "^4.3.1",
     "bootstrap-vue": "^2.17.3",
-    "core-js": "^2.6.5",
+    "core-js": "^3.6.5",
+    "font-awesome": "^4.7.0",
     "jquery": "^3.4.1",
     "lz-string": "^1.4.4",
     "popper.js": "^1.15.0",
     "v-runtime-template": "^1.10.0",
-    "vue": "^2.6.10",
-    "vue-router": "^3.0.7",
+    "vue": "^2.6.12",
+    "vue-router": "^3.4.6",
     "vue-slider-component": "^3.2.9",
     "vue2-daterange-picker": "^0.6.1",
     "vuedraggable": "^2.24.3",
     "vuex": "^3.1.1"
   },
   "devDependencies": {
-    "@types/jest": "^23.1.4",
-    "@vue/cli-plugin-babel": "^3.9.0",
+    "@types/jest": "^26.0.14",
+    "@typescript-eslint/eslint-plugin": "^4.4.0",
+    "@typescript-eslint/parser": "^4.4.0",
+    "@vue/cli-plugin-babel": "~4.5.7",
     "@vue/cli-plugin-e2e-nightwatch": "^3.9.0",
-    "@vue/cli-plugin-eslint": "^3.9.0",
+    "@vue/cli-plugin-eslint": "~4.5.7",
+    "@vue/cli-plugin-router": "~4.5.7",
     "@vue/cli-plugin-typescript": "^3.9.0",
-    "@vue/cli-plugin-unit-jest": "^3.9.0",
+    "@vue/cli-plugin-unit-jest": "^4.5.11",
     "@vue/cli-service": "^3.9.0",
     "@vue/eslint-config-standard": "^4.0.0",
-    "@vue/eslint-config-typescript": "^4.0.0",
-    "@vue/test-utils": "1.0.0-beta.29",
+    "@vue/eslint-config-typescript": "^6.0.0",
+    "@vue/test-utils": "^1.1.0",
     "babel-core": "7.0.0-bridge.0",
     "babel-eslint": "^10.0.1",
+    "babel-jest": "^26.6.3",
     "banner-webpack-plugin": "^0.2.3",
     "chromedriver": "^76.0.1",
-    "eslint": "^5.16.0",
-    "eslint-plugin-vue": "^5.0.0",
+    "eslint": "^7.10.0",
+    "eslint-plugin-vue": "^7.0.1",
     "geckodriver": "^1.21.1",
+    "node-sass": "^4.12.0",
+    "sass-loader": "^10.0.2",
     "ts-jest": "^23.0.0",
-    "typescript": "^3.6.3",
-    "vue-template-compiler": "^2.6.10"
+    "typescript": "~4.0.3",
+    "vue-template-compiler": "^2.6.12"
   },
   "resolutions": {
     "handlebars": "4.5.0"
