+++ packages/data-explorer/public/index.html
@@ -7,6 +7,13 @@
     <link rel="icon" href="<%= BASE_URL %>favicon.ico">
     <title>data-explorer</title>
     <script>
+    window.__INITIAL_STATE__ = {
+      baseUrl: '/',
+      lng: 'en',
+      fallbackLng: 'en',
+      dataExplorerBaseUrl: '/sys_Language/en'
+    }
+
       // Detect IE 10 and IE 11
     document.addEventListener('DOMContentLoaded', function () {
       if (/Trident\/|MSIE/.test(window.navigator.userAgent)) {
