+++ packages/data-explorer/src/routes.ts
@@ -1,13 +1,26 @@
 import MainView from './views/MainView.vue'
+import DataRowEdit from '@/components/DataRowEdit.vue'
 
 const routes = [
   {
-    path: '/:entity/:view',
-    component: MainView,
-    name: 'main-view'
+    path: '/explorer/:entity/create',
+    props: true,
+    name: 'de-create',
+    component: DataRowEdit
   },
   {
-    path: '/:entity?',
+    path: '/explorer/:entity/:dataRowId/edit',
+    props: true,
+    name: 'de-edit',
+    component: DataRowEdit
+  },
+  {
+    path: '/explorer/:entity/:view',
+    component: MainView,
+    name: 'de-view'
+  },
+  {
+    path: '/',
     redirect: to => {
       let entity = 'sys_ts_DataExplorerEntitySettings'
       if (to.params.entity) {
@@ -15,7 +28,7 @@ const routes = [
       }
 
       return {
-        name: 'main-view',
+        name: 'de-view',
         params: { entity, view: 'CardView' },
         query: { page: '1', size: '20', hideSidebar: 'false' }
       }
