+++ packages/data-explorer/tests/unit/setup.js
@@ -1,0 +1,5 @@
+const VueTestUtils = require('@vue/test-utils')
+VueTestUtils.config.mocks.$t = key => key
+
+// Distinguish test-specific code paths.
+global.IS_TEST = true
