+++ packages/data-explorer/tests/unit/views/ToolbarView.spec.ts
@@ -6,60 +6,67 @@ import ApplicationState from '@/types/ApplicationState'
 import { Sort } from '@/types/Sort'
 import bootstrapVue from 'bootstrap-vue'
 
-describe('ToolbarView.vue', () => {
-  const $t = (msg: any) => msg
+let actions:any
+let mutations: any
+let getters: any
+let wrapper:any
+let store: any
+let state: ApplicationState
+
+const $t = (msg: any) => msg
+const mocks: any = {
+  $route: { params: {}, query: {} },
+  $router: { push: jest.fn() },
+  $t
+}
+
+function getWrapper () {
   const localVue = createLocalVue()
   localVue.use(Vuex)
   localVue.use(bootstrapVue)
   localVue.filter('i18n', $t)
-  let actions: any
-  let store: any
-  let state: ApplicationState
-  let mutations: any
-  let getters: any
-  let directives: any
-  const mocks: any = {
-    $route: { params: {}, query: {} },
-    $router: {
-      push: jest.fn()
-    },
-    $t
+
+  state = mockState()
+  state.tableName = 'root_hospital_patients'
+
+  actions = {
+    downloadResources: jest.fn()
   }
 
-  beforeEach(() => {
-    state = mockState()
-    state.tableName = 'root_hospital_patients'
+  mutations = {
+    setDataDisplayLayout: jest.fn(),
+    setFilterSelection: jest.fn(),
+    setSearchText: jest.fn()
+  }
 
-    actions = {
-      downloadResources: jest.fn()
-    }
+  getters = {
+    hasEditRights: jest.fn(),
+    hasAddRights: jest.fn(),
+    hasEditSettingsRights: jest.fn(),
+    compressedRouteFilter: jest.fn()
+  }
 
-    mutations = {
-      setDataDisplayLayout: jest.fn(),
-      setFilterSelection: jest.fn(),
-      setSearchText: jest.fn()
-    }
+  getters.hasAddRights.mockReturnValue(true)
+  store = new Vuex.Store({ actions, state, mutations, getters })
 
-    getters = {
-      hasEditRights: jest.fn(),
-      hasAddRights: jest.fn(),
-      hasEditSettingsRights: jest.fn(),
-      compressedRouteFilter: jest.fn()
-    }
-
-    store = new Vuex.Store({
-      actions, state, mutations, getters
-    })
-
-    directives = {
+  return shallowMount(ToolbarView, {
+    store,
+    localVue,
+    directives: {
       'b-tooltip': () => {}
-    }
+    },
+    mocks,
+    stubs: ['router-link']
+  })
+}
+
+describe('ToolbarView.vue', () => {
+  beforeEach(() => {
+    wrapper = getWrapper()
   })
 
   describe('when the search text model updates', () => {
-    let wrapper
     beforeEach(() => {
-      wrapper = shallowMount(ToolbarView, { store, localVue, directives, mocks })
       wrapper.setData({ searchText: 'test' })
     })
 
@@ -70,56 +77,48 @@ describe('ToolbarView.vue', () => {
   })
 
   it('exists', () => {
-    const wrapper = shallowMount(ToolbarView, { store, localVue, directives, mocks })
     expect(wrapper.exists()).toBeTruthy()
   })
 
   it('can change to table layout', () => {
     state.dataDisplayLayout = 'CardView'
-    store = new Vuex.Store({
-      state, mutations, getters
-    })
-    const wrapper = shallowMount(ToolbarView, { store, localVue, directives, mocks })
+
     wrapper.find('button.table-layout').trigger('click')
     expect(mocks.$router.push).toHaveBeenCalledWith({
-      name: 'main-view',
+      name: 'de-view',
       params: { view: 'TableView' },
       query: {}
     })
   })
 
-  it('can change to card layout', () => {
-    state.dataDisplayLayout = 'TableView'
-    store = new Vuex.Store({
-      state, mutations, getters
-    })
-    const wrapper = shallowMount(ToolbarView, { store, localVue, directives, mocks })
+  it('can change to card layout', async () => {
+    store.state.dataDisplayLayout = 'TableView'
+    await wrapper.vm.$nextTick()
     wrapper.find('button.card-layout').trigger('click')
     expect(mocks.$router.push).toHaveBeenCalledWith({
-      name: 'main-view',
+      name: 'de-view',
       params: { view: 'CardView' },
       query: {}
     })
   })
 
   describe('add row button', () => {
-    beforeEach(() => getters.hasAddRights.mockReturnValueOnce(true))
+    beforeEach(async () => {
+      getters.hasAddRights.mockReturnValue(true)
+    })
 
-    it('should render the add button as a link to the data-row-edit', () => {
-      const wrapper = shallowMount(ToolbarView, { store, localVue, directives, mocks })
-      expect(wrapper.find('.btn-toolbar > div > a.add-row').attributes().href).toEqual('/plugin/data-row-edit/root_hospital_patients')
+    it('should render the add button as a link to the data-row-edit', async () => {
+      expect(wrapper.find('.add-row').exists()).toBe(true)
     })
 
     it('should not be shown in shoppingcart mode', () => {
       store.state.showSelected = true
-      const wrapper = shallowMount(ToolbarView, { store, localVue, directives, mocks })
       expect(wrapper.find('.btn-toolbar > a').exists()).toBe(false)
     })
 
     it('should not show the button without edit rights', () => {
       store.state.showSelected = true
       getters.hasEditRights.mockReturnValueOnce(false)
-      const wrapper = shallowMount(ToolbarView, { store, localVue, directives, mocks })
       expect(wrapper.find('.btn-toolbar > a').exists()).toBe(false)
     })
 
@@ -129,7 +128,6 @@ describe('ToolbarView.vue', () => {
   describe('searchString value is set', () => {
     it('should persist the value in the store', () => {
       store.state.showSelected = true
-      const wrapper = shallowMount(ToolbarView, { store, localVue, directives, mocks })
       wrapper.setData({ searchText: 'demo' })
       expect(mutations.setSearchText).toHaveBeenCalled()
     })
@@ -138,7 +136,6 @@ describe('ToolbarView.vue', () => {
   describe('downloadData', () => {
     it('download action called', async () => {
       store.state.tableMeta = { id: '123', attributes: [] }
-      const wrapper = shallowMount(ToolbarView, { store, localVue, directives, mocks })
       // @ts-ignore
       await wrapper.vm.downloadData()
       expect(actions.downloadResources).toHaveBeenCalledWith(expect.anything(), [{ id: '123', type: 'ENTITY_TYPE' }])
@@ -147,7 +144,6 @@ describe('ToolbarView.vue', () => {
 
   describe('sortOptions', () => {
     it('should return an empty list while metaData is empty', () => {
-      const wrapper = shallowMount(ToolbarView, { store, localVue, directives, mocks })
       // @ts-ignore
       expect(wrapper.vm.sortOptions).toEqual([])
     })
@@ -161,7 +157,6 @@ describe('ToolbarView.vue', () => {
           { visible: true, id: '3', name: 'c', type: 'compound' }
         ]
       }
-      const wrapper = shallowMount(ToolbarView, { store, localVue, directives, mocks })
       // @ts-ignore
       expect(wrapper.vm.sortOptions).toEqual([{ id: '1', name: 'a' }])
     })
@@ -169,14 +164,13 @@ describe('ToolbarView.vue', () => {
 
   describe('handleSortSelectChange', () => {
     it('should update the route query', async () => {
-      const wrapper = shallowMount(ToolbarView, { store, localVue, directives, mocks })
       const sort: Sort = {
         sortColumnName: 'stars',
         isSortOrderReversed: true
       }
       // @ts-ignore
       await wrapper.vm.handleSortSelectChange(sort)
-      expect(mocks.$router.push).toHaveBeenCalledWith({ name: 'main-view', query: { sort: '-stars' } })
+      expect(mocks.$router.push).toHaveBeenCalledWith({ name: 'de-view', query: { sort: '-stars' } })
     })
   })
 })
