--- packages/data-row-edit/package.json
@@ -1,50 +1,0 @@
-{
-  "name": "@molgenis-ui/data-row-edit",
-  "version": "5.2.0",
-  "license": "LGPL-3.0",
-  "description": "Data row edit",
-  "author": "connor <c.stroomberg@rug.nl>",
-  "main": "dist/index.html",
-  "files": [
-    "dist",
-    "src"
-  ],
-  "publishConfig": {
-    "access": "public"
-  },
-  "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "unit": "vue-cli-service test:unit",
-    "lint": "vue-cli-service lint",
-    "version": "yarn"
-  },
-  "dependencies": {
-    "@molgenis/molgenis-api-client": "^3.1.3",
-    "@molgenis/molgenis-i18n-js": "^0.2.1",
-    "@molgenis/molgenis-ui-form": "^5.3.0",
-    "core-js": "^3.6.5",
-    "font-awesome": "^4.7.0",
-    "vue": "^2.6.12",
-    "vue-router": "^3.4.6"
-  },
-  "devDependencies": {
-    "@types/jest": "^26.0.14",
-    "@typescript-eslint/eslint-plugin": "^4.4.0",
-    "@typescript-eslint/parser": "^4.4.0",
-    "@vue/cli-plugin-babel": "~4.5.7",
-    "@vue/cli-plugin-eslint": "~4.5.7",
-    "@vue/cli-plugin-router": "~4.5.7",
-    "@vue/cli-plugin-typescript": "~4.5.7",
-    "@vue/cli-plugin-unit-jest": "~4.5.7",
-    "@vue/cli-service": "~4.5.7",
-    "@vue/eslint-config-typescript": "^6.0.0",
-    "@vue/test-utils": "^1.1.0",
-    "eslint": "^7.10.0",
-    "eslint-plugin-vue": "^7.0.1",
-    "node-sass": "^4.12.0",
-    "sass-loader": "^10.0.2",
-    "typescript": "~4.0.3",
-    "vue-template-compiler": "^2.6.12"
-  }
-}
