--- packages/data-row-edit/public/index.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width,initial-scale=1.0">
-    <link rel="shortcut icon" href="https://avatars1.githubusercontent.com/u/1688158?v=4&s=280"/>
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css"
-          integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css">
-    <title>MOLGENIS - Data row edit</title>
-  </head>
-  <body>
-    <noscript>
-      <strong>We're sorry but <%= htmlWebpackPlugin.options.title %> doesn't work properly without JavaScript enabled. Please enable it to continue.</strong>
-    </noscript>
-    <div id="data-row-edit-plugin"></div>
-    <!-- built files will be auto injected -->
-    <script>
-      window.__INITIAL_STATE__ = {
-        baseUrl: '/',
-        lng: 'en',
-        fallbackLng: 'en',
-        dataExplorerBaseUrl: '/sys_Language/en'
-      }
-    </script>
-  </body>
-</html>
