--- packages/data-row-edit/src/App.vue
@@ -1,6 +1,0 @@
-<template>
-  <div id="data-row-edit-plugin">
-    <router-view/>
-  </div>
-</template>
-
